unify default behaviors
