+++ cob_default_robot_behavior/setup.py
@@ -1,0 +1,11 @@
+#!/usr/bin/env python
+
+from distutils.core import setup
+from catkin_pkg.python_setup import generate_distutils_setup
+
+d = generate_distutils_setup(
+   packages=['cob_default_robot_behavior'],
+   package_dir={'': 'src'}
+)
+
+setup(**d)
